Fix Roadmap ID assignment and remove debug log + misc changes
The changes fix the incorrect Roadmap Id assignment logic when in non-creation mode. Before, `pageId` was being used, which refers to the urlPath instead of the correct `roadmap.id`. Therefore, it's replaced by the correct `roadmap.id`. Additionally, a debug log statement is removed from the `roadmap-state-protocols.ts` as it's not necessary for the production code.