Redirect to '/explore' if roadmap not found
This commit fixes an issue where the application would return 500 when navigating to a non-existing roadmap.
Now, users are redirected to the '/explore' page in case of typo in the URL or deleted content. The code now checks if a roadmap is present after fetching from the API. If not post-validation, it performs a redirection.