Add 'feeling lucky' functionality to Explore feature
This commit adds a 'Feeling Lucky' feature to the Explore component. Instead of manually setting the explore query, the 'Feeling Lucky' button now makes a call to the newly created 'fetchFeelingLucky' function in the 'roadmap-card-data' API wrapper. This function fetches a random roadmap and redirects the user to it. This adds an element of surprise and spontaneity to the user experience.