Add floating effect to objects in Home component
To improve the visual effect in the Home component, new attributes (sinOffset, cosOffset) were added to objects, allowing them to have a floating effect. This floating effect is achieved by adjusting the y-position of objects using a sine function. The value of 'time' is also introduced and updated in each animation frame for this purpose. The enhancement improves the user experience by giving the page a more dynamic and alive feeling.