Refactor home components and extract helpers
The ScrollingElement and Home components from the frontend's home directory are refactored for better readability and maintainability - states and effects are rearranged, classNames are reformatted, key values for maps are changed to uuids for better uniqueness and their functions are extracted into a helper file. The helper functions are exported for better testing and maintainability. In addition, a new dependency, "@types/uuid", is added to the package.json and package-lock.json to handle unique id generation for key values in maps.