Set mouse position to screen center by default
The changes in this commit are to address the issue where mouse positions were not properly initialized. When the mouse position coordinates (x, y) were set to (0,0), it caused the animation to jump to from 0,0 to coordinates of the first mouse move. By initializing the mouse position to the screen center regardless of the actual mouse position, we can dampen the jump of the animation code.