Add aria-labels to home page buttons and move NotificationProvider
The changes in this commit involve improving accessibility.

Firstly, 'aria-label' attributes are added to the buttons on the home page in order to make them more accessible for screen readers. This attribute provides a text description for these buttons, ensuring their purpose could be understood even when images fail to load or for visually impaired users.

Secondly, the NotificationProvider is repositioned within Base.astro file for better encapsulation, making the body the parent element instead of html. This change is intended to isolate notifications to the body of the page to better control its placement and avoid any potential overlap with other elements.