Set default values for text, node and component properties
Updated multiple components and properties to have default values to prevent any possibility of null or undefined value issues. This ensures no unexpected crashing when trying to access a certain property within the node, text or component. Defaults consider the basic and most common configuration of a node, such as dimensions or opacity. Additionally enhanced DraggableInput to handle defaultValue and validate its input, improving the robustness of input handling.