Updated NodeContextMenu to handle outside clicks
Added useRef and useEffect to handle outside and child click events in NodeContextMenu component. This is to improve usability by allowing the editing of status for child nodes. Adjustments made to the NodeContextMenu appearance properties and also repositioned the context menu to follow the cursor more accurately in NodeRendererForeign. Added node progress as a prop to NodeContextMenu for better status management.