Mobile Roadmap view
This commit factors out the 'Recenter' button from the CoordsRoadmapElement component into its own RecenterButton component. This allows for reusability.

In addition, changes were made to the styling of several components to better cater to various screen sizes. This includes modifying gap sizes, text sizes, and the positioning of several elements. The use of 'md:hidden' and 'hidden md:block' tailwind classes help to make the interface more responsive.

Moreover, the addition of 'NavigoLearn' to metadata.title in RoadmapLayout.astro ensures a more consistent and professional appearance in the browser title bar. Similarly, the body's font has been set to 'kanit-text' in several components for visual consistency across the application.