Add support for context menu on mobile Safari
Added a new hook useContextMenuOrLongPress to handle both right click events and long press events on mobile devices. This was necessary to provide an equivalent of the 'right-clicking' option for mobile users. Also, refactoring was performed on the useIsMobile hook for a cleaner implementation by extracting a re-usable function checkIsMobile.

Additionally, the NodeRendererClassic component was updated in order to utilise the new hook and now provides instructions to either long-tap or right-click on a node to modify its progress status, accordingly to the platform in use.